Compile pendingList and DisplayTeamMembers into Management

+++ src/pages/api/team/[id].js
@@ -1,5 +1,5 @@
 import connectDB from "../../../backend/lib/connectDB";
-import { StatusCodes } from "http-status-codes";
+import {StatusCodes} from "http-status-codes";
 import Teams from "../../../backend/models/team";
 import Course from "../../../backend/models/course";
 import User from "../../../backend/models/user";
@@ -19,11 +19,11 @@ export default async function handler(req, res) {
                      * Update Information
                      */
                     const {
-                        query: { id }
+                        query: {id}
                     } = req
                     //Update Team information
                     if (req.body.newName && req.body.newCourse && req.body.newDescription) {
-                        const courseName = await Course.findOne({ name: req.body.newCourse }, "_id").lean()
+                        const courseName = await Course.findOne({name: req.body.newCourse}, "_id").lean()
                         const courseId = courseName._id.toString()
 
                         const newTeam = {
@@ -33,18 +33,18 @@ export default async function handler(req, res) {
                         }
 
                         const team = await Teams.findByIdAndUpdate(
-                            id, newTeam, { new: true, runValidators: true }
+                            id, newTeam, {new: true, runValidators: true}
                         )
                         if (!team) {
-                            new Error.json({ message: "Team not found" })
+                            new Error.json({message: "Team not found"})
                         }
 
-                        return res.status(StatusCodes.OK).json({ message: "Your account updated" })
+                        return res.status(StatusCodes.OK).json({message: "Your account updated"})
                     } else if (req.body.userId) { //Update Team members
                         const team = await Teams.findById(id)
                         team.userID.push(req.body.userId)
                         await team.save()
-                        return res.status(StatusCodes.OK).json({ message: "Welcome" })
+                        return res.status(StatusCodes.OK).json({message: "Welcome"})
                     }
                     break
                 } catch (e) {
@@ -57,32 +57,37 @@ export default async function handler(req, res) {
                      * Delete User from Team
                      */
                     const {
-                        query: { id }
+                        query: {id}
                     } = req
 
-                    //Delete main
-                    const team = await Teams.findByIdAndDelete(id)
+                    if (req.body) {
+                        console.log(req.body)
+                        res.status(StatusCodes.OK).json(req.body);
+                    } else {
+                        //Delete main
+                        const team = await Teams.findByIdAndDelete(id)
 
-                    //Delete reference
-                    await User.findByIdAndUpdate(team.userID, { $pull: { post_id: team._id } })
-                    const lists = await List.find({ _id: { $in: team.listID } }, "task_id").lean()
+                        //Delete reference
+                        await User.findByIdAndUpdate(team.userID, {$pull: {post_id: team._id}})
+                        const lists = await List.find({_id: {$in: team.listID}}, "task_id").lean()
 
-                    await List.deleteMany({ _id: { $in: team.listID } })
+                        await List.deleteMany({_id: {$in: team.listID}})
 
-                    lists.map(async (list) => {
-                        await Task.deleteMany({ _id: { $in: list.task_id } })
-                    })
+                        lists.map(async (list) => {
+                            await Task.deleteMany({_id: {$in: list.task_id}})
+                        })
 
 
-                    return res.status(StatusCodes.OK).json({ message: "Deleted" })
+                        return res.status(StatusCodes.OK).json({message: "Deleted"})
+                    }
                 } catch (e) {
                     return e
                 }
             }
         }
     } catch
-    (Error) {
-        return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ Error })
+        (Error) {
+        return res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({Error})
     }
 
 }
